JS
inputfield
New JS operators (&&-- This is the Logical AND operator.

This operator is commonly used to check if all supplied values are truthy.

!=
This is the Inequality operator.

This operator returns true if each side represents a different value

Unlike the strict inequality operator (!==), the value's type isn't considered. For example, the number 2 is considered equivalent to the string "2", so the expression 2 != '2' returns false. This is generally a bad thing, and it's recommended to use the strict inequality operator instead.)